requirements.txt
langchain>=0.1.0
langchain-cohere>=0.1.0
langchain-experimental>=0.0.50
pandas>=2.0.0
python-dotenv>=1.0.0
ragas>=0.1.0
datasets>=2.14.0
numpy>=1.24.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
openevals>=0.1.2